+++ src/Commands/DomainActionMakeCommand.php
@@ -2,9 +2,7 @@
 
 namespace Lunarstorm\LaravelDDD\Commands;
 
-use Symfony\Component\Console\Input\InputArgument;
-
-class MakeAction extends DomainGeneratorCommand
+class DomainActionMakeCommand extends DomainGeneratorCommand
 {
     protected $name = 'ddd:action';
 
@@ -17,27 +15,9 @@ class MakeAction extends DomainGeneratorCommand
 
     protected $type = 'Action';
 
-    protected function getArguments()
-    {
-        return [
-            ...parent::getArguments(),
-
-            new InputArgument(
-                'name',
-                InputArgument::REQUIRED,
-                'The name of the Action',
-            ),
-        ];
-    }
-
     protected function getStub()
     {
         return $this->resolveStubPath('action.php.stub');
-    }
-
-    protected function getRelativeDomainNamespace(): string
-    {
-        return config('ddd.namespaces.actions', 'Actions');
     }
 
     protected function preparePlaceholders(): array