+++ src/Commands/DomainDtoMakeCommand.php
@@ -2,9 +2,7 @@
 
 namespace Lunarstorm\LaravelDDD\Commands;
 
-use Symfony\Component\Console\Input\InputArgument;
-
-class MakeDTO extends DomainGeneratorCommand
+class DomainDtoMakeCommand extends DomainGeneratorCommand
 {
     protected $name = 'ddd:dto';
 
@@ -17,19 +15,6 @@ class MakeDTO extends DomainGeneratorCommand
 
     protected $type = 'Data Transfer Object';
 
-    protected function getArguments()
-    {
-        return [
-            ...parent::getArguments(),
-
-            new InputArgument(
-                'name',
-                InputArgument::REQUIRED,
-                'The name of the DTO',
-            ),
-        ];
-    }
-
     protected function getStub()
     {
         return $this->resolveStubPath('dto.php.stub');
@@ -37,7 +22,7 @@ class MakeDTO extends DomainGeneratorCommand
 
     protected function getRelativeDomainNamespace(): string
     {
-        return config('ddd.namespaces.data_transfer_objects', 'Data');
+        return config('ddd.namespaces.data_transfer_object', 'Data');
     }
 
     protected function preparePlaceholders(): array