+++ src/Commands/DomainEnumMakeCommand.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Lunarstorm\LaravelDDD\Commands;
+
+use Illuminate\Foundation\Console\EnumMakeCommand;
+use Lunarstorm\LaravelDDD\Commands\Concerns\ResolvesDomainFromInput;
+
+class DomainEnumMakeCommand extends EnumMakeCommand
+{
+    use ResolvesDomainFromInput;
+
+    protected $name = 'ddd:enum';
+}