+++ src/Commands/DomainEventMakeCommand.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Lunarstorm\LaravelDDD\Commands;
+
+use Illuminate\Foundation\Console\EventMakeCommand;
+use Lunarstorm\LaravelDDD\Commands\Concerns\ResolvesDomainFromInput;
+
+class DomainEventMakeCommand extends EventMakeCommand
+{
+    use ResolvesDomainFromInput;
+
+    protected $name = 'ddd:event';
+}