+++ src/Commands/DomainRuleMakeCommand.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Lunarstorm\LaravelDDD\Commands;
+
+use Illuminate\Foundation\Console\RuleMakeCommand;
+use Lunarstorm\LaravelDDD\Commands\Concerns\ResolvesDomainFromInput;
+
+class DomainRuleMakeCommand extends RuleMakeCommand
+{
+    use ResolvesDomainFromInput;
+
+    protected $name = 'ddd:rule';
+}