+++ tests/Fixtures/Enums/Feature.php
@@ -6,6 +6,7 @@ enum Feature: string
 {
     case PromptForMissingInput = '9.49.0';
     case IncludeFilepathInGeneratorCommandOutput = '9.32.0';
+    case LaravelPromptsPackage = '10.17';
 
     public function exists(): bool
     {