+++ tests/Generator/MakeBaseModelTest.php
@@ -14,7 +14,7 @@ it('can generate domain base model', function ($domainPath, $domainRoot) {
     $relativePath = implode('/', [
         $domainPath,
         $domain,
-        config('ddd.namespaces.models'),
+        config('ddd.namespaces.model'),
         "{$modelName}.php",
     ]);
 
@@ -26,7 +26,7 @@ it('can generate domain base model', function ($domainPath, $domainRoot) {
 
     expect(file_exists($expectedPath))->toBeFalse();
 
-    Artisan::call("ddd:base-model {$domain} {$modelName}");
+    Artisan::call("ddd:base-model {$domain}:{$modelName}");
 
     expect(Artisan::output())->when(
         Feature::IncludeFilepathInGeneratorCommandOutput->exists(),
@@ -38,14 +38,8 @@ it('can generate domain base model', function ($domainPath, $domainRoot) {
     $expectedNamespace = implode('\\', [
         $domainRoot,
         $domain,
-        config('ddd.namespaces.models'),
+        config('ddd.namespaces.model'),
     ]);
 
     expect(file_get_contents($expectedPath))->toContain("namespace {$expectedNamespace};");
 })->with('domainPaths');
-
-it('shows meaningful hints when prompting for missing input', function () {
-    $this->artisan('ddd:base-model')
-        ->expectsQuestion('What is the domain?', 'Shared')
-        ->assertExitCode(0);
-});