+++ tests/Generator/MakeBaseViewModelTest.php
@@ -2,7 +2,6 @@
 
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\Config;
-use Lunarstorm\LaravelDDD\Tests\Fixtures\Enums\Feature;
 
 it('can generate base view model', function ($domainPath, $domainRoot) {
     Config::set('ddd.domain_path', $domainPath);
@@ -14,7 +13,7 @@ it('can generate base view model', function ($domainPath, $domainRoot) {
     $relativePath = implode('/', [
         $domainPath,
         $domain,
-        config('ddd.namespaces.view_models'),
+        config('ddd.namespaces.view_model'),
         "{$className}.php",
     ]);
 
@@ -26,26 +25,17 @@ it('can generate base view model', function ($domainPath, $domainRoot) {
 
     expect(file_exists($expectedPath))->toBeFalse();
 
-    Artisan::call("ddd:base-view-model {$domain} {$className}");
+    Artisan::call("ddd:base-view-model {$domain}:{$className}");
 
-    expect(Artisan::output())->when(
-        Feature::IncludeFilepathInGeneratorCommandOutput->exists(),
-        fn ($output) => $output->toContainFilepath($relativePath),
-    );
+    expect(Artisan::output())->toContainFilepath($relativePath);
 
     expect(file_exists($expectedPath))->toBeTrue();
 
     $expectedNamespace = implode('\\', [
         $domainRoot,
         $domain,
-        config('ddd.namespaces.view_models'),
+        config('ddd.namespaces.view_model'),
     ]);
 
     expect(file_get_contents($expectedPath))->toContain("namespace {$expectedNamespace};");
 })->with('domainPaths');
-
-it('shows meaningful hints when prompting for missing input', function () {
-    $this->artisan('ddd:base-view-model')
-        ->expectsQuestion('What is the domain?', 'Shared')
-        ->assertExitCode(0);
-});