+++ tests/Generator/MakeViewModelTest.php
@@ -3,7 +3,6 @@
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Str;
-use Lunarstorm\LaravelDDD\Tests\Fixtures\Enums\Feature;
 
 it('can generate view models', function ($domainPath, $domainRoot) {
     Config::set('ddd.domain_path', $domainPath);
@@ -15,7 +14,7 @@ it('can generate view models', function ($domainPath, $domainRoot) {
     $relativePath = implode('/', [
         $domainPath,
         $domain,
-        config('ddd.namespaces.view_models'),
+        config('ddd.namespaces.view_model'),
         "{$viewModelName}.php",
     ]);
 
@@ -27,19 +26,16 @@ it('can generate view models', function ($domainPath, $domainRoot) {
 
     expect(file_exists($expectedPath))->toBeFalse();
 
-    Artisan::call("ddd:view-model {$domain} {$viewModelName}");
+    Artisan::call("ddd:view-model {$domain}:{$viewModelName}");
 
-    expect(Artisan::output())->when(
-        Feature::IncludeFilepathInGeneratorCommandOutput->exists(),
-        fn ($output) => $output->toContainFilepath($relativePath),
-    );
+    expect(Artisan::output())->toContainFilepath($relativePath);
 
     expect(file_exists($expectedPath))->toBeTrue();
 
     $expectedNamespace = implode('\\', [
         $domainRoot,
         $domain,
-        config('ddd.namespaces.view_models'),
+        config('ddd.namespaces.view_model'),
     ]);
 
     expect(file_get_contents($expectedPath))->toContain("namespace {$expectedNamespace};");
@@ -51,11 +47,11 @@ it('normalizes generated view model to pascal case', function ($given, $normaliz
     $expectedPath = base_path(implode('/', [
         config('ddd.domain_path'),
         $domain,
-        config('ddd.namespaces.view_models'),
+        config('ddd.namespaces.view_model'),
         "{$normalized}.php",
     ]));
 
-    Artisan::call("ddd:view-model {$domain} {$given}");
+    Artisan::call("ddd:view-model {$domain}:{$given}");
 
     expect(file_exists($expectedPath))->toBeTrue();
 })->with('makeViewModelInputs');
@@ -67,7 +63,7 @@ it('generates the base view model if needed', function () {
     $expectedPath = base_path(implode('/', [
         config('ddd.domain_path'),
         $domain,
-        config('ddd.namespaces.view_models'),
+        config('ddd.namespaces.view_model'),
         "{$className}.php",
     ]));
 
@@ -86,14 +82,7 @@ it('generates the base view model if needed', function () {
 
     expect(file_exists($expectedBaseViewModelPath))->toBeFalse();
 
-    Artisan::call("ddd:view-model {$domain} {$className}");
+    Artisan::call("ddd:view-model {$domain}:{$className}");
 
     expect(file_exists($expectedBaseViewModelPath))->toBeTrue();
-});
-
-it('shows meaningful hints when prompting for missing input', function () {
-    $this->artisan('ddd:view-model')
-        ->expectsQuestion('What is the domain?', 'Utility')
-        ->expectsQuestion('What should the view model be named?', 'Belt')
-        ->assertExitCode(0);
 });