+++ tests/Support/DomainResolverTest.php
@@ -1,0 +1,25 @@
+<?php
+
+use Lunarstorm\LaravelDDD\Support\DomainResolver;
+
+beforeEach(function () {
+    $this->artisan('ddd:model', [
+        'name' => 'Invoice',
+        '--domain' => 'Invoicing',
+    ]);
+
+    $this->artisan('ddd:dto', [
+        'name' => 'CustomerProfile',
+        '--domain' => 'Customer',
+    ]);
+
+    $this->expectedDomains = [
+        'Customer',
+        'Invoicing',
+        'Shared',
+    ];
+});
+
+it('can get the current domains', function () {
+    expect(DomainResolver::domainChoices())->toEqualCanonicalizing($this->expectedDomains);
+});